api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
pymongo==4.6.1
pydantic==2.5.3
python-dotenv==1.0.0
google-auth==2.27.0
PyJWT==2.8.0

# OG Image Generation
Pillow==10.2.0
aiohttp==3.9.1
google-cloud-storage==2.14.0
